update toolbox dependency
This will truncate buttons with ... if the text gets too long.
Add an explicit max-width to the button to enable this feature.

Note: this requires a workaround of {`Choose ${device}`} because
of the way I set up toolbox to truncate text nodes. Previously,
it truncated "choose" and "device" separately. This is a bug
I should fix in toolbox.