Grid View: remove recursion
The recursion didn't really provide any benefit, it merely was an
n-1 passthrough until it got the end of the path.

Remove the recursion and render directly the thing we want
given the path. There are two cases: render a grid and render
a preview of a single device. Now we decide up front which
to render based on the path.