Device preview: Click to select device
Now there are so many more places you can click in order to select the
device! We didn't see a reason not to provide the biggest reasonable
area. A screen reader shouldn't be affected by the extra ways to trigger
a selection; this will just affect people using the mouse to navigate
through the options.