useAsync: Silence react-hooks/exhaustive-deps
There's nothing the `useAsync()` wrapper around `useEffect()` can do to
address the warnings, so we may as well silence them. One of the
downsides of `useAsync()` is that it hides the `useEffect()` from the
linter, so it can't warn us when the dependencies being passed into
`useAsync()` are incomplete.

Case in point, we weren't passing in `fetchChildren` as a dependency;
now we do. It doesn't need to be wrapped in `useCallback()` because we
expect that to be done up the tree if it's necessary, but it isn't
currently because it's defined outside of a component.