DevicePicker: Focus search input on mount
This is convenient for all users as they can start typing right away
(while still being able to navigate with keys), and it helps screen
readers, which are otherwise left with their focus wherever it was in
the DOM, far away from the new device picker tree.

As it happens, we already had a reference to the mounted input.