Test App: Read from query parameters
When the easy query parameter API is available (should be all modern
browsers save IE, of course), set the initial device to the value of the
'd' query parameter and use grid view if the 'g' query parameter is
present (because I like the column view more and think it should be the
default).

This required modifying `DevicePicker.tsx` to export `DevicePicker`
separately from making it the default export.