Device preview: Click only image, title, summary
The image makes this a bit tricky because it's a background image on a
div that spans the available space. That is, there is no image element
to make clickable/hoverable. We don't want to make the container
a specific or maximum size because then we'd lose our centering. So we
give each of the image, title, and text a maximum width on the
assumption that it doesn't make visual sense to give the summary a
maximum width if the other two elements can be vastly wider.

This means that a little white space around the image is clickable (up
to the 30em maximum width, corresponding to where the summary wraps,
when it wraps). I don't think it's worth further changing the way that
we render the preview image in order to make that space unclickable.