useAsync: Try another clarification
After some discussion, Andrew and I settled on this approach to try to
make the code explain itself. We found it difficult to capture our
intent and the complicated interactions with `useEffect` in a couple
names, but hopefully this is a step in the right direction. The point of
it all is just that we don't want to use the results of `fn` if the
dependencies have changed since we called it (including implicit
dependencies like the underlying component being mounted, and not
rerendered).