Device Picker: Setting the Stage
Device picker now has two children, the algolia and the Hierarchical
category based viewers. These rotate based on whether there is an active
search query or if a non-leaf (a category, EG "GPS") is selected.
algolia does not store any information about whether a hit is a leaf or
not. That is where the findAllLeaves function comes in. It recursively
looks through the whole tree and if any node is a leaf it records it
for later use. The category or search query are stored in one common
union variable: context.