Algolia Device Picker: wire into the system
Apart from several style changes, the bulk of this commit is adding onClick
functionality to each hit. This is determined by the leaf data
structure, if a hit is a leaf, onClick it will select it. Otherwise, it
resets the context variable of DevicePicker, and renders the
Hierarchical Device picker at that category. Side note: the
connectSearchBox and connectHits functions make it possible to use our
own styled components with their tech.